@mbutrovich 2018-08-09
Current master on my laptop:
```
BM_SimpleInsert/repeats:10/real_time_mean              859.092MB/s   17.1818M items/s
BM_SimpleInsert/repeats:10/real_time_median            852.427MB/s   17.0485M items/s
```
This PR:
```
BM_SimpleInsert/repeats:10/real_time_mean              934.931MB/s   18.6986M items/s
BM_SimpleInsert/repeats:10/real_time_median            937.454MB/s   18.7491M items/s
```
We'll take 8-10% any time. Nice job!