@mbutrovich 2018-08-09
So it looks like the merge of master into this branch broke things. I tried an interactive rebase from your first commit but you changed concurrent_bitmap_test in both of your PRs in the last day, and I want to preserve your intended changes from both so I'll have to wait for you to look at it.